requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
base58==2.1.1
    # via mm-solana (pyproject.toml)
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   mm-solana (pyproject.toml)
    #   click-aliases
click-aliases==1.0.4
    # via mm-solana (pyproject.toml)
construct==2.10.68
    # via construct-typing
construct-typing==0.5.6
    # via solana
cryptography==43.0.0
    # via mm-std
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via solana
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via mm-solana (pyproject.toml)
jsonalias==0.1.1
    # via solders
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mm-std==0.1.0
    # via mm-solana (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via mm-std
pydantic-core==2.20.1
    # via pydantic
pydash==8.0.3
    # via mm-std
pygments==2.18.0
    # via rich
pysocks==1.7.1
    # via mm-std
python-dateutil==2.9.0.post0
    # via mm-std
python-dotenv==1.0.1
    # via mm-std
pyyaml==6.0.1
    # via
    #   mm-solana (pyproject.toml)
    #   mm-std
requests==2.32.3
    # via mm-std
rich==13.7.1
    # via mm-std
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
solana==0.34.2
    # via mm-solana (pyproject.toml)
solders==0.21.0
    # via solana
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
    #   pydash
    #   solana
    #   solders
urllib3==2.2.2
    # via requests
websockets==11.0.3
    # via solana
